requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
class-registry==2.1.2
cryptography==42.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-phonenumber-field==7.2.0
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filters==1.3.2
filters-django==1.0.5
firebase-admin==6.5.0
frozenlist==1.4.1
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googlemaps==4.10.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jsonfield==3.1.0
jwcrypto==1.5.6
mailtrap==2.0.1
MarkupPy==1.14
msgpack==1.0.8
multidict==6.0.5
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.27
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.2.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
python-jwt==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
swapper==1.3.0
tablib==3.5.0
twilio==9.0.0
types-pkg-resources==0.1.3
typing_extensions==4.12.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
